pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.47.0",
    "pyside6>=6.10.0",
    "python-docx>=1.2.0",
]
